WSL/servicios/personas-api/requirements.txt
flask
mysql-connector-python
prometheus_client